Never drop update-feed issues from queue

+++ db.go
@@ -389,14 +389,15 @@ func (c *DBClient) RetryQueuedIssue(ctx context.Context, key string) error {
 	defer tx.Rollback()
 
 	var failedCount int
-	query := `SELECT failed_count FROM sync_queue WHERE issue_key = $1 FOR UPDATE`
-	err = tx.QueryRowContext(ctx, query, key).Scan(&failedCount)
+	var priority int
+	query := `SELECT failed_count, priority FROM sync_queue WHERE issue_key = $1 FOR UPDATE`
+	err = tx.QueryRowContext(ctx, query, key).Scan(&failedCount, &priority)
 	if err != nil {
 		return err
 	}
 
 	failedCount += 1
-	if failedCount > 4 {
+	if failedCount > 4 && priority < 10 {
 		_, err = tx.ExecContext(ctx, `DELETE FROM sync_queue WHERE issue_key = $1`, key)
 	} else {
 		// Delay will be: 5m, 25m, 2h5m, 10h25m
